Move Sensor objects and save_readouts to ControllerGUITab
Since the GUI frame actually controls the sensors, we'll do the saving in the frame object instead
